review changes to make app function
